Addition of self-cal loop for pipeline (loop 3)
Loop of phase self-cals that checks if the stations are coherent and applies solution. It will triple the solution interval if stations are not coherent. It will break if the solution interval exceeds 1 hour. What still needs to be added is if only some stations are coherent once that time frame is reached.