fix for phase solutions with no polarization axis
The script could not handle phase000 solutions without polarization axis (so from a scalar type solve).  soltab.getValuesIter(returnAxes=['pol', 'ant', 'freq', 'time'] returned an error in that case. The rest of the script was already correct catching this case checking via if 'pol' in soltab.getAxesNames(): but the first check was overlooked. I added it and verified it worked.